+++ src/app/d2-loadout-builder/generated-sets/GeneratedSetButtons.tsx
@@ -1,10 +1,12 @@
 import { t } from 'i18next';
+import copy from 'fast-copy';
 import * as React from 'react';
 import * as _ from 'lodash';
 import { DimStore } from '../../inventory/store-types';
 import { dimLoadoutService, Loadout, LoadoutClass } from '../../loadout/loadout.service';
 import { ArmorSet } from '../types';
 import { newLoadout } from '../../loadout/loadout-utils';
+import { AppIcon, powerIndicatorIcon } from '../../shell/icons';
 
 /**
  * Renders the Create Loadout and Equip Items buttons for each generated set
@@ -25,12 +27,7 @@ export default function GeneratedSetButtons({
 
   // Automatically equip items for this generated set to the active store
   const equipItems = () => {
-    const loadout: Loadout = createLoadout(store.class, set);
-
-    _.each(loadout.items, (val) => {
-      val[0].equipped = true;
-    });
-
+    const loadout = createLoadout(store.class, set);
     return dimLoadoutService.applyLoadout(store, loadout, true);
   };
 
@@ -42,7 +39,9 @@ export default function GeneratedSetButtons({
       <button className="dim-button equip-button" onClick={equipItems}>
         {t('LoadoutBuilder.EquipItems', { name: store.name })}
       </button>
-      <span className="light">{set.power / set.armor.length}</span>
+      <span className="light">
+        <AppIcon icon={powerIndicatorIcon} /> {set.power / set.armor.length}
+      </span>
     </div>
   );
 }
@@ -51,13 +50,21 @@ export default function GeneratedSetButtons({
  * Create a Loadout object, used for equipping or creating a new saved loadout
  */
 function createLoadout(classType: DimStore['class'], set: ArmorSet): Loadout {
-  const loadout = newLoadout(t('Loadouts.AppliedAuto'), {
-    helmet: [set.armor[0]],
-    gauntlets: [set.armor[1]],
-    chest: [set.armor[2]],
-    leg: [set.armor[3]],
-    classitem: [set.armor[4]]
-  });
+  const loadout = newLoadout(
+    t('Loadouts.AppliedAuto'),
+    copy({
+      helmet: [set.armor[0]],
+      gauntlets: [set.armor[1]],
+      chest: [set.armor[2]],
+      leg: [set.armor[3]],
+      classitem: [set.armor[4]]
+    })
+  );
   loadout.classType = LoadoutClass[classType];
+
+  _.each(loadout.items, (val) => {
+    val[0].equipped = true;
+  });
+
   return loadout;
 }