+++ docs/CHANGELOG.md
@@ -1,10 +1,14 @@
 # Next
 
+* We no longer support searching D1 vendor items.
+
+# 4.74.0 (2018-10-14)
+
 * Added negative search. Prefix any search term with `-` and it will match the opposite.
 * Added `perk:"* **"` seach filter to match any keywords against perks on an item
 * Added some missing `stat:`
 * Lock and unlock items matching your current search from the same menu you use for tagging them.
-* We no longer support searching D1 vendor items.
+* Updated icons across the app.d
 
 # 4.73.0 (2018-10-07)